pyproject.toml
[tool.poetry]
name = "polkaalert"
version = "0.1.0"
description = ""
authors = ["Marcos Antonio <marcossk7@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
python-telegram-bot = "^21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
